Extract Netv6Addr into its own module
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
