+++ src/lib.rs
@@ -326,8 +326,10 @@ pub use mask::*;
 
 mod netaddr;
 mod netv4addr;
+mod netv6addr;
 pub use netaddr::*;
 pub use netv4addr::*;
+pub use netv6addr::*;
 
 mod netaddr_error;
 pub use netaddr_error::*;
