+++ src/netv6addr.rs
@@ -1,0 +1,23 @@
+use crate::mask::Mask;
+use std::net::Ipv6Addr;
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash, Ord)]
+pub struct Netv6Addr {
+	mask: Ipv6Addr,
+	addr: Ipv6Addr,
+}
+
+impl Netv6Addr {
+	pub(crate) fn mask(&self) -> &Ipv6Addr {
+		&self.mask
+	}
+
+	pub(crate) fn addr(&self) -> &Ipv6Addr {
+		&self.addr
+	}
+
+	pub fn new(addr: Ipv6Addr, mask: Ipv6Addr) -> Self {
+		let addr = addr.mask(&mask);
+		Self { addr, mask }
+	}
+}
